[GLib] Remove usage of API::DownloadClient::legacyDidCancel https://bugs.webkit.org/show_bug.cgi?id=289724
Reviewed by NOBODY (OOPS!).

This is mostly a revert of b73ca668112d75dd0975cd8568740eb65f3ff103

An implementation of legacyDidCancel() was added in that commit, however
that method is intended to be removed.

Since we aren't actually relying on the client recieving didCancel over the
proxy there is no reason to use it.

Instead to fix the original bug that commit intended just accept that
didFail or didFinish can be called in any order since cancellation is async.

* Source/WebKit/UIProcess/API/glib/WebKitDownload.cpp:
(webkit_download_cancel):
* Source/WebKit/UIProcess/API/glib/WebKitDownloadClient.cpp: